moved nb_letter to new file count.c
